requirements.txt
Flask==2.2.2
Pillow==9.2.0
face_recognition==1.3.0
opencv-python==4.5.5.64
